+++ src/routes/(app)/(protected)/search/user/[username]/+page.server.ts
@@ -25,8 +25,8 @@ export const load: PageServerLoad = async (event: ServerLoadEvent) => {
 		username
 	);
 	const data: UserDataWithPosts = {
-		userData: userData,
-		postData: postData
+		userData,
+		postData
 	};
 	return data;
 };
